Handle recursive union types
checkAmbiguousMemberTypes might trigger calls to createOrReuseUnion when
fetching the fields of a given type. However, if graphQLType is not
defined yet, it will try to create it instead of returning the
previously created one. In order to reuse graphQLType without creating a
new one, we need to call checkAmbiguousMemberTypes **after** assigning
the graphQLType to the current definition.

Signed-off-by: Simon Meyffret <smeyffret@gmail.com>